requirements.txt
anthropic==0.7.5
faiss-cpu==1.7.4
huggingface-hub==0.19.4
langchain==0.0.341
langsmith==0.0.66
openai==1.3.5
pydantic==2.5.2
PyYAML==6.0.1
sentence-transformers==2.2.2
streamlit==1.28.2
torch==2.1.1 -f https://download.pytorch.org/whl/cpu/torch_stable.html
